change the way we deploy contract in dev and test env
This is how current flow looks:
1) create authority account
2) fund it
3) deploy RootChain.sol using authority account
4) use it to mine operator blocks

Reason to do it this way is that `geth --dev` will not mine blocks
unless we give him transaction to mine. This means we need an account to
send such transaction and we can't use authority account to do that.

Whole thing is enabled by helper in DevHelper

Other changes:
* run geth with personal api enabled on rpc interface
* incorporate BlockQueue into happy path integration test
* minor rewrite of eth.ex