mad scenario #1 identifies two problems with the model
1) there is a error in a way split translates into amount1/amount2
between :exec and next_state(.., :exec, ..). This is the real issue.
2) some block mining from scenario is being dropped invalidating the scenario
(marked as #1 and #2). While regardless of number of blocks mined
second spend is invalid (second tx does not exist no matter which blknum you are
looking into), seeing that command was not updated to reflect world
where blocks were not mined is sad.