feat: allow to challenge the unchallengable unchallenged exit (WIP)
...in situation where watcher realized the invalidity and lateness at the same time

Before we stopped "in the middle" of applying of a block, if we realized that there's an unchallenged exit.
This lead to the in-memory state of the ledger (State.Core) be updated, but not the DB.
As a consequence, on couldn't get the proper challenging info, which is inconvenient and weird.

Now an unchallenged exit stops only the process of applying new blocks allowing the already started transactions to complete.

This is enough, because the unchallenged exit byzantine condition is "soft" anyway.