reset rpc_counter and use cassette instead
When we use the `@tag fixtures:` functionality, it calls out to
`DevGeth` and etc ... which will send out a request from Ethereumex to
ensure that geth is loaded. This whoever causes the request id to
increment and causes all the cassettes to not much when you remove the
fixtures call. We explicitly reset the counter before we start the
actual interactions in the tests.