fix: tidy the behavior of finalizing piggybacked outputs
Now inclusion and position of said piggybacks (and as a consequence - the need to exit UTXOs from `OMG.State`) is now decided in `ExitProcessor`.

This leads to the change of behavior, that canonical but not-included IFE txs will be able to exit their outputs.

Slightly related to #994, this allows to treat #994 as a feature request, not a bug.