chore: add more Tesla middleware overriding for a more stable client on load tests (#1381)
* chore: use Jason instead of Poison

From @Ino's knowledge, Jason is faster than Poison.

ref: https://github.com/phoenixframework/phoenix/issues/2693#issuecomment-388881650

* chore: add retry to tesla to enhance stability of load test

To avoid transient error breaking the laod test.

* chore: consistent formatter

Running mix format in elixir-omg/ and elixir-omg/priv/perf would result in different format changes.
This make them consistent on the line_lenght of 120.

* chore: add Logger to Tesla middleware

Load test runnings long. We need a way to know what has happened to debug and fix.

ref: https://hexdocs.pm/tesla/Tesla.Middleware.Logger.html#content

* chore: change user-agents header to be clear it is from test

In case we are tracking logs in server side, we want to differentiate where it comes from.

* chore: use connection pool and setup timeout of Tesla

This setting enforces Tesla client to use :per_pool as the connection pool. The pool would be initiated on the start
of application. pool_size and max_connection can be set by config file.

ref:
- hackney connection pool: https://github.com/benoitc/hackney#use-the-default-pool
- tesla opt: https://hexdocs.pm/tesla/Tesla.Middleware.Opts.html
- tesla adapter: https://github.com/teamon/tesla/wiki/0.x-to-1.0-Migration-Guide#adapter-options-need-to-be-wrapped-in-adapter-key
- elixir Application: https://hexdocs.pm/elixir/Application.html

* chore: update README for the changes and config setup

* fix: remove impl true

seems like there is no "behavior" for this...