feat: test with payment v2 transfer
Tested by hacking the watcher info /transaction.create to generate payment v2 tx instead.
And then the tx would be signed and send normally. Tested with the cabbage transaction tests.

This commit instead of providing new module for payment, it reuses most of the part instead of the function
that generate the payment struct. This commit introduces Payment.Builder to build different payment tx.

Probably not the most clean way of handling things but the least change required to allow cabbage test to work.