test: cabbage test for payment v2 (#1578)
* feat: test with payment v2 transfer

Tested by hacking the watcher info /transaction.create to generate payment v2 tx instead.
And then the tx would be signed and send normally. Tested with the cabbage transaction tests.

This commit instead of providing new module for payment, it reuses most of the part instead of the function
that generate the payment struct. This commit introduces Payment.Builder to build different payment tx.

Probably not the most clean way of handling things but the least change required to allow cabbage test to work.

* feat: allow /transaction.create to pass in tx_type

add tx_type as a param of 'order'. If not passed, default to 1 for backward competibility.

* test: cabbage test for payment v2 transfer

* fix: cleanup

linter, dialyzer, remove branching on tx type when building payment tx.

* refactor: use Alice and Bob instead

Simplify the test code and make it easier to read

* fix: credo complain single pipe