test: payment v2 with standard exit (#1589)
* feat: multiple exit processors (#1573)

* feat: start multiple exit processors

* chore: use snapshot support payment v2

... and run mix format

* fix: setup exit_games for test fixture

* fix: dispatch return in the format of {:ok, []}

* fix: make watcher integ test to work

* fix: unit test for EthEventAggregator

1. fix the unit tests, add address to log result
2. change how we bind contract address to log, use the pre-decoded address data

* fix: patch payment_exit_game back into contract_addr

* fix: wrong filter_events mapping

* fix: linter, dializer, remove tmp logging

* fix: use snapwhot with payment v2

Previous snapshot was wrong.
Using this one:
https://github.com/omisego-images/docker-elixir-omg/pull/43#issuecomment-620469853

* refactor: restrain from using ++

Use [a|b] as best practice

* chore: use Task.async_stream instad of sync looping

* refactor: move private functions to the end

* fix: map after Task.async_stream

Co-authored-by: Thibault Denizet <thibault@omisego.co>

* feat: test with payment v2 transfer

Tested by hacking the watcher info /transaction.create to generate payment v2 tx instead.
And then the tx would be signed and send normally. Tested with the cabbage transaction tests.

This commit instead of providing new module for payment, it reuses most of the part instead of the function
that generate the payment struct. This commit introduces Payment.Builder to build different payment tx.

Probably not the most clean way of handling things but the least change required to allow cabbage test to work.

* feat: allow /transaction.create to pass in tx_type

add tx_type as a param of 'order'. If not passed, default to 1 for backward competibility.

* test: cabbage test for payment v2 transfer

* fix: cleanup

linter, dialyzer, remove branching on tx type when building payment tx.

* refactor: use Alice and Bob instead

Simplify the test code and make it easier to read

* fix: credo complain single pipe

* test: payment v2 with standard exit

* fix: compiler warning me :(

* refactor: remove commented codes

* fix: linter

* style: simpler code

* refactor: parameterize users

Co-authored-by: Thibault Denizet <thibault@omisego.co>