feat: add new metadata to Datadog traces
Because API responses are always 200, even if there's an error message
in the body of the response, it's hard to get metrics from Datadog about
 the health of services.

What you see in Datadog traces:
* The service name (default 'web' for Phoenix applications) is the name
of the OMG application.
* Traces are tagged with application version number
* For API responses which error, the trace is flagged red as an error,
the error code is marked on the trace and the error message is found
in the tags.